requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
appdirs==1.4.3
astor==0.8.1
async-timeout==3.0.1
atari-py==0.2.6
attrs==20.3.0
blessings==1.7
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
colorful==0.5.4
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
dm-tree==0.1.5
filelock==3.0.12
future==0.18.2
gast==0.2.2
google-api-core==1.26.3
google-auth==1.28.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.36.1
gym==0.18.0
h5py==3.2.1
hiredis==2.0.0
html5lib==1.0.1
idna==2.10
importlib-metadata==3.10.0
ipaddr==2.2.0
joblib==1.0.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lockfile==0.12.2
lz4==3.1.3
Markdown==3.3.4
msgpack==1.0.2
multidict==5.1.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
opencensus==0.7.12
opencensus-context==0.1.2
opencv-python==4.5.1.48
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
prometheus-client==0.9.0
protobuf==3.15.6
psutil==5.8.0
py==1.10.0
py-spy==0.3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pymetasploit3==1.0.3
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
ray==1.2.0
redis==3.5.3
requests==2.25.1
retry==0.9.2
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.14.0
sklearn==0.0
tabulate==0.8.9
tensorboard==1.15.0
tensorboardX==2.1
tensorflow-estimator==1.15.1
# tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.1
typing-extensions==3.7.4.3
urllib3==1.26.3
webencodings==0.5.1pyt
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
